Update final project structure
