+++ tuandat-360review/public/index.html
@@ -25,203 +25,12 @@
   <!-- CSS Just for demo purpose, don't include it in your project -->
   <link href="./assets/demo/demo.css" rel="stylesheet" />
 
-  <style type="text/css">
-    body {
-      height: 100%;
-      margin: 0;
-      padding: 0;
-      width: 100%
-    }
-
-    .ipl-progress-indicator.available {
-      opacity: 0
-    }
-
-    .ipl-progress-indicator {
-      background-color: #f5f5f5;
-      width: 100%;
-      height: 100%;
-      position: fixed;
-      opacity: 1;
-      pointer-events: none;
-      -webkit-transition: opacity cubic-bezier(.4, 0, .2, 1) 436ms;
-      -moz-transition: opacity cubic-bezier(.4, 0, .2, 1) 436ms;
-      transition: opacity cubic-bezier(.4, 0, .2, 1) 436ms;
-      z-index: 9999
-    }
-
-    .insp-logo-frame {
-      display: -webkit-flex;
-      display: -moz-flex;
-      display: flex;
-      -webkit-flex-direction: column;
-      -moz-flex-direction: column;
-      flex-direction: column;
-      -webkit-justify-content: center;
-      -moz-justify-content: center;
-      justify-content: center;
-      -webkit-animation: fadein 436ms;
-      -moz-animation: fadein 436ms;
-      animation: fadein 436ms;
-      height: 98%
-    }
-
-    .insp-logo-frame-img {
-      width: 112px;
-      height: 112px;
-      -webkit-align-self: center;
-      -moz-align-self: center;
-      align-self: center;
-      border-radius: 50%
-    }
-
-    .ipl-progress-indicator-head {
-      background-color: #c6dafc;
-      height: 4px;
-      overflow: hidden;
-      position: relative
-    }
-
-    .ipl-progress-indicator .first-indicator,
-    .ipl-progress-indicator .second-indicator {
-      background-color: #056d8b;
-      bottom: 0;
-      left: 0;
-      right: 0;
-      top: 0;
-      position: absolute;
-      -webkit-transform-origin: left center;
-      -moz-transform-origin: left center;
-      transform-origin: left center;
-      -webkit-transform: scaleX(0);
-      -moz-transform: scaleX(0);
-      transform: scaleX(0)
-    }
-
-    .ipl-progress-indicator .first-indicator {
-      -webkit-animation: first-indicator 2s linear infinite;
-      -moz-animation: first-indicator 2s linear infinite;
-      animation: first-indicator 2s linear infinite
-    }
-
-    .ipl-progress-indicator .second-indicator {
-      -webkit-animation: second-indicator 2s linear infinite;
-      -moz-animation: second-indicator 2s linear infinite;
-      animation: second-indicator 2s linear infinite
-    }
-
-    .ipl-progress-indicator .insp-logo {
-      animation: App-logo-spin infinite 20s linear;
-      border-radius: 50%;
-      -webkit-align-self: center;
-      -moz-align-self: center;
-      align-self: center
-    }
-
-    @keyframes App-logo-spin {
-      from {
-        transform: rotate(0)
-      }
-
-      to {
-        transform: rotate(360deg)
-      }
-    }
-
-    @-webkit-keyframes fadein {
-      from {
-        opacity: 0
-      }
-
-      to {
-        opacity: 1
-      }
-    }
-
-    @-moz-keyframes fadein {
-      from {
-        opacity: 0
-      }
-
-      to {
-        opacity: 1
-      }
-    }
-
-    @keyframes fadein {
-      from {
-        opacity: 0
-      }
-
-      to {
-        opacity: 1
-      }
-    }
-
-    @keyframes first-indicator {
-      0% {
-        transform: translate(0) scaleX(0)
-      }
-
-      25% {
-        transform: translate(0) scaleX(.5)
-      }
-
-      50% {
-        transform: translate(25%) scaleX(.75)
-      }
-
-      75% {
-        transform: translate(100%) scaleX(0)
-      }
-
-      100% {
-        transform: translate(100%) scaleX(0)
-      }
-    }
-
-    @keyframes second-indicator {
-      0% {
-        transform: translate(0) scaleX(0)
-      }
-
-      60% {
-        transform: translate(0) scaleX(0)
-      }
-
-      80% {
-        transform: translate(0) scaleX(.6)
-      }
-
-      100% {
-        transform: translate(100%) scaleX(.1)
-      }
-    }
-  </style>
 </head>
 
 <body>
   <noscript>You need to enable JavaScript to run this app.</noscript>
-  <!-- <div class="ipl-progress-indicator" id="ipl-progress-indicator">
-    <div class="ipl-progress-indicator-head">
-      <div class="first-indicator"></div>
-      <div class="second-indicator"></div>
-    </div>
-    <div class="insp-logo-frame">
-      <svg width="112" class="insp-logo-frame-img" height="112" xmlns="http://www.w3.org/2000/svg"
-        viewBox="0 0 841.9 595.3">
-        <g fill="#61DAFB">
-          <path
-            d="M666.3 296.5c0-32.5-40.7-63.3-103.1-82.4 14.4-63.6 8-114.2-20.2-130.4-6.5-3.8-14.1-5.6-22.4-5.6v22.3c4.6 0 8.3.9 11.4 2.6 13.6 7.8 19.5 37.5 14.9 75.7-1.1 9.4-2.9 19.3-5.1 29.4-19.6-4.8-41-8.5-63.5-10.9-13.5-18.5-27.5-35.3-41.6-50 32.6-30.3 63.2-46.9 84-46.9V78c-27.5 0-63.5 19.6-99.9 53.6-36.4-33.8-72.4-53.2-99.9-53.2v22.3c20.7 0 51.4 16.5 84 46.6-14 14.7-28 31.4-41.3 49.9-22.6 2.4-44 6.1-63.6 11-2.3-10-4-19.7-5.2-29-4.7-38.2 1.1-67.9 14.6-75.8 3-1.8 6.9-2.6 11.5-2.6V78.5c-8.4 0-16 1.8-22.6 5.6-28.1 16.2-34.4 66.7-19.9 130.1-62.2 19.2-102.7 49.9-102.7 82.3 0 32.5 40.7 63.3 103.1 82.4-14.4 63.6-8 114.2 20.2 130.4 6.5 3.8 14.1 5.6 22.5 5.6 27.5 0 63.5-19.6 99.9-53.6 36.4 33.8 72.4 53.2 99.9 53.2 8.4 0 16-1.8 22.6-5.6 28.1-16.2 34.4-66.7 19.9-130.1 62-19.1 102.5-49.9 102.5-82.3zm-130.2-66.7c-3.7 12.9-8.3 26.2-13.5 39.5-4.1-8-8.4-16-13.1-24-4.6-8-9.5-15.8-14.4-23.4 14.2 2.1 27.9 4.7 41 7.9zm-45.8 106.5c-7.8 13.5-15.8 26.3-24.1 38.2-14.9 1.3-30 2-45.2 2-15.1 0-30.2-.7-45-1.9-8.3-11.9-16.4-24.6-24.2-38-7.6-13.1-14.5-26.4-20.8-39.8 6.2-13.4 13.2-26.8 20.7-39.9 7.8-13.5 15.8-26.3 24.1-38.2 14.9-1.3 30-2 45.2-2 15.1 0 30.2.7 45 1.9 8.3 11.9 16.4 24.6 24.2 38 7.6 13.1 14.5 26.4 20.8 39.8-6.3 13.4-13.2 26.8-20.7 39.9zm32.3-13c5.4 13.4 10 26.8 13.8 39.8-13.1 3.2-26.9 5.9-41.2 8 4.9-7.7 9.8-15.6 14.4-23.7 4.6-8 8.9-16.1 13-24.1zM421.2 430c-9.3-9.6-18.6-20.3-27.8-32 9 .4 18.2.7 27.5.7 9.4 0 18.7-.2 27.8-.7-9 11.7-18.3 22.4-27.5 32zm-74.4-58.9c-14.2-2.1-27.9-4.7-41-7.9 3.7-12.9 8.3-26.2 13.5-39.5 4.1 8 8.4 16 13.1 24 4.7 8 9.5 15.8 14.4 23.4zM420.7 163c9.3 9.6 18.6 20.3 27.8 32-9-.4-18.2-.7-27.5-.7-9.4 0-18.7.2-27.8.7 9-11.7 18.3-22.4 27.5-32zm-74 58.9c-4.9 7.7-9.8 15.6-14.4 23.7-4.6 8-8.9 16-13 24-5.4-13.4-10-26.8-13.8-39.8 13.1-3.1 26.9-5.8 41.2-7.9zm-90.5 125.2c-35.4-15.1-58.3-34.9-58.3-50.6 0-15.7 22.9-35.6 58.3-50.6 8.6-3.7 18-7 27.7-10.1 5.7 19.6 13.2 40 22.5 60.9-9.2 20.8-16.6 41.1-22.2 60.6-9.9-3.1-19.3-6.5-28-10.2zM310 490c-13.6-7.8-19.5-37.5-14.9-75.7 1.1-9.4 2.9-19.3 5.1-29.4 19.6 4.8 41 8.5 63.5 10.9 13.5 18.5 27.5 35.3 41.6 50-32.6 30.3-63.2 46.9-84 46.9-4.5-.1-8.3-1-11.3-2.7zm237.2-76.2c4.7 38.2-1.1 67.9-14.6 75.8-3 1.8-6.9 2.6-11.5 2.6-20.7 0-51.4-16.5-84-46.6 14-14.7 28-31.4 41.3-49.9 22.6-2.4 44-6.1 63.6-11 2.3 10.1 4.1 19.8 5.2 29.1zm38.5-66.7c-8.6 3.7-18 7-27.7 10.1-5.7-19.6-13.2-40-22.5-60.9 9.2-20.8 16.6-41.1 22.2-60.6 9.9 3.1 19.3 6.5 28.1 10.2 35.4 15.1 58.3 34.9 58.3 50.6-.1 15.7-23 35.6-58.4 50.6zM320.8 78.4z" />
-          <circle cx="420.9" cy="296.5" r="45.7" />
-          <path d="M520.5 78.1z" />
-        </g>
-      </svg>
-    </div>
-  </div> -->
-
+  
   <div id="root"></div>
-
 
 
   <!--   Core JS Files   -->
