+++ tuandat-360review/src/components/Atoms/FooterDetail/index.js
