+++ tuandat-360review/src/components/Molecules/container/assign/index.js
@@ -1,0 +1,13 @@
+import React, { Component } from 'react';
+
+import AssignDetail from '../../../Atoms/AssignDetail/';
+
+class Assign extends Component {
+  render() {
+    return (
+      <AssignDetail />
+    )
+  }
+}
+
+export default Assign;
