+++ tuandat-360review/src/components/Molecules/container/footer/index.js
@@ -1,0 +1,8 @@
+import React from 'react';
+import Detail from '../../../Atoms/FooterDetail/';
+
+const footer = () => (
+    <Detail />
+)
+
+export default footer;
