+++ tuandat-360review/src/components/Molecules/container/question/create.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import QuestionDetail from '../../../Atoms/QuestionDetail/';
+
+class Question extends Component {
+  render() {
+    return (
+      <QuestionDetail />
+    )
+  }
+}
+
+export default Question;
