+++ tuandat-360review/src/components/Molecules/container/question/list.js
@@ -1,0 +1,12 @@
+import React, { Component } from 'react';
+import ListDetail from '../../../Atoms/QuestionDetail/list';
+
+class List extends Component {
+  render() {
+    return (
+      <ListDetail />
+    )
+  }
+}
+
+export default List;
