+++ tuandat-360review/src/components/Molecules/container/report/index.js
@@ -1,0 +1,21 @@
+import React, { Component } from 'react';
+import Table from '../../../Atoms/ReportDetail';
+
+import { connect } from 'react-redux';
+import * as taskActions from '../../../../actions/task.js';
+
+class Report extends Component {
+
+  componentDidMount() {
+    const { fetchLoadPage } = this.props;
+    fetchLoadPage(); // from taskActions
+  }
+
+  render() {
+    return (
+      <Table />
+    )
+  }
+}
+
+export default connect(null, { ...taskActions })(Report);
