--- tuandat-360review/src/components/ReportDetail/pagination/index.js
@@ -1,26 +1,0 @@
-import React from 'react';
-
-const pagination = () => (
-  <div className="mt-4">
-    <div className="pull-left">
-      <p>Hiển thị 4 trong tổng số 20</p>
-    </div>
-    <nav className="pull-right" aria-label="Page navigation">
-      <ul className="pagination justify-content-center">
-        <li className="page-item disabled">
-          <a className="page-link" href="#" tabIndex="-1">Previous</a>
-        </li>
-        <li className="page-item active"><a className="page-link" href="#">1</a></li>
-        <li className="page-item"><a className="page-link" href="#">2</a></li>
-        <li className="page-item"><a className="page-link" href="#">3</a></li>
-        <li className="page-item"><a className="page-link" href="#">4</a></li>
-        <li className="page-item"><a className="page-link" href="#">5</a></li>
-        <li className="page-item">
-          <a className="page-link" href="#">Next</a>
-        </li>
-      </ul>
-    </nav>
-  </div>
-)
-
-export default pagination;
