--- tuandat-360review/src/components/container/footer/index.js
@@ -1,8 +1,0 @@
-import React from 'react';
-import Detail from '../../FooterDetail/';
-
-const footer = () => (
-    <Detail />
-)
-
-export default footer;
