--- tuandat-360review/src/components/menu/index.js
@@ -1,57 +1,0 @@
-import React, { Component } from 'react';
-import { Link } from 'react-router-dom';
-
-
-//import Aux from '../../hoc/Aux.js'
-import { withTranslation } from 'react-i18next';
-
-class Menu extends Component {
-  render() {
-    const { t } = this.props;
-
-    return (
-      <div className="sidebar" data-color="white" data-active-color="danger">
-        {/* Logo */}
-        <div className="logo">
-          <a href="#" className="logo-normal">
-            <div className="logo-image-small">
-              <img src="/assets/img/hybrid-logo.jpg" alt="Logo" />
-            </div>
-          </a>
-        </div>
-        {/* Menu Item */}
-        <div className="sidebar-wrapper">
-          <ul className="nav">
-            <li className="active ">
-              <Link to="/evaluation">
-                <i className="nc-icon nc-bulb-63"></i>
-                <p>{t('menu.evaluate')}</p>
-              </Link>
-            </li>
-            <li>
-              <a href="#">
-                <i className="nc-icon nc-glasses-2"></i>
-                <p>{t('menu.question')}</p>
-              </a>
-            </li>
-            <li>
-              <a href="#">
-                <i className="nc-icon nc-bullet-list-67"></i>
-                <p>{t('menu.assign')}</p>
-              </a>
-            </li>
-            <li>
-              <Link to="/report">
-                <i className="nc-icon nc-paper"></i>
-                <p>{t('menu.report')}</p>
-              </Link>
-            </li>
-          </ul>
-        </div>
-
-      </div>
-    )
-  }
-}
-
-export default withTranslation()(Menu);
